This program does vehicles detection and tracking.
This program uses yolo11m-obb pretrained model for vehicle detections, and YOLO in-built BoT-SORT algorithm with Veri-WILD ReID pretrained model. It draws the trail path with distinguishable colors for visualization of the tracking. 

The BoT-SORT.yaml config file has manually tuned configurations, and the path to the VERI-WILD ReID pretrained model, and the ReID configurations.
The links to the BoT-SORT algorithm source code of ultralytics, and the Fast ReID's VERI-WILD ReID pretrained model has been provided in the .yaml file as comments at the top.